+++ R/gg_boxplot.R
@@ -63,7 +63,7 @@ gg_boxplot <- function(data,
                        subtitle = NULL,
                        subtitle_wrap = 80,
                        x_balance = FALSE,
-                       x_labels = waiver(),
+                       x_labels = NULL,
                        x_na = TRUE,
                        x_pretty_n = 6,
                        x_expand = NULL,
@@ -191,6 +191,7 @@ gg_boxplot <- function(data,
     x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_var_vctr), max(x_var_vctr))
     if(is.null(x_expand)) x_expand <- c(0, 0)
+    if(is.null(x_labels)) x_labels <- waiver()
     
     if(mobile == TRUE) {
       x_breaks <- x_limits
@@ -228,6 +229,7 @@ gg_boxplot <- function(data,
   }
   else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
+    if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
     plot <- plot +
       scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -377,7 +379,7 @@ gg_boxplot_col <- function(data,
                            subtitle = NULL,
                            subtitle_wrap = 80,
                            x_balance = FALSE,
-                           x_labels = waiver(),
+                           x_labels = NULL,
                            x_pretty_n = 6,
                            x_expand = NULL,
                            x_na = TRUE,
@@ -532,6 +534,7 @@ gg_boxplot_col <- function(data,
     x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_var_vctr), max(x_var_vctr))
     if(is.null(x_expand)) x_expand <- c(0, 0)
+    if(is.null(x_labels)) x_labels <- waiver()
     
     if(mobile == TRUE) {
       x_breaks <- x_limits
@@ -569,6 +572,7 @@ gg_boxplot_col <- function(data,
   }
   else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
+    if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
     plot <- plot +
       scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -724,7 +728,7 @@ gg_boxplot_facet <- function(data,
                              subtitle_wrap = 80,
                              x_balance = FALSE,
                              x_expand = NULL,
-                             x_labels = waiver(),
+                             x_labels = NULL,
                              x_na = TRUE,
                              x_pretty_n = 6,
                              x_rev = FALSE,
@@ -871,6 +875,7 @@ gg_boxplot_facet <- function(data,
         x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = FALSE)
         x_limits <- c(min(x_var_vctr), max(x_var_vctr))
         if(is.null(x_expand)) x_expand <- c(0, 0)
+        if(is.null(x_labels)) x_labels <- waiver()
       }
       
       if (is.numeric(x_var_vctr)) {
@@ -903,6 +908,7 @@ gg_boxplot_facet <- function(data,
       }
       else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
         if(is.null(x_expand)) x_expand <- waiver()
+        if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
         
         plot <- plot +
           scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -1053,7 +1059,7 @@ gg_boxplot_col_facet <- function(data,
                                  subtitle = NULL,
                                  subtitle_wrap = 80,
                                  x_balance = FALSE,
-                                 x_labels = waiver(),
+                                 x_labels = NULL,
                                  x_na = TRUE,
                                  x_pretty_n = 6,
                                  x_expand = NULL,
@@ -1222,6 +1228,7 @@ gg_boxplot_col_facet <- function(data,
       x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = FALSE)
       x_limits <- c(min(x_var_vctr), max(x_var_vctr))
       if(is.null(x_expand)) x_expand <- c(0, 0)
+      if(is.null(x_labels)) x_labels <- waiver()
     }
     
     if (is.numeric(x_var_vctr)) {
@@ -1254,6 +1261,7 @@ gg_boxplot_col_facet <- function(data,
     }
     else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
       if(is.null(x_expand)) x_expand <- waiver()
+      if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
       
       plot <- plot +
         scale_x_discrete(expand = x_expand, labels = x_labels)