+++ R/gg_line.R
@@ -61,7 +61,7 @@ gg_line <- function(data,
                     subtitle_wrap = 80,
                     x_balance = FALSE,
                     x_expand = NULL,
-                    x_labels = waiver(),
+                    x_labels = NULL,
                     x_na = TRUE,
                     x_pretty_n = 6,
                     x_rev = FALSE,
@@ -148,6 +148,7 @@ gg_line <- function(data,
     x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_breaks), max(x_breaks))
     if(is.null(x_expand)) x_expand <- c(0, 0)
+    if(is.null(x_labels)) x_labels <- waiver()
     
     if(mobile == TRUE) {
       x_breaks <- x_limits
@@ -188,6 +189,7 @@ gg_line <- function(data,
   }
   else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
+    if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
     plot <- plot +
       scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -323,7 +325,7 @@ gg_line_col <- function(data,
                         subtitle_wrap = 80,
                         x_balance = FALSE,
                         x_expand = NULL,
-                        x_labels = waiver(),
+                        x_labels = NULL,
                         x_na = TRUE,
                         x_pretty_n = 6,
                         x_rev = FALSE,
@@ -435,6 +437,7 @@ gg_line_col <- function(data,
     x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_breaks), max(x_breaks))
     if(is.null(x_expand)) x_expand <- c(0, 0)
+    if(is.null(x_labels)) x_labels <- waiver()
     
     if(mobile == TRUE) {
       x_breaks <- x_limits
@@ -475,6 +478,7 @@ gg_line_col <- function(data,
   }
   else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
+    if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
     plot <- plot +
       scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -614,7 +618,7 @@ gg_line_facet <- function(data,
                           subtitle_wrap = 80,
                           x_balance = FALSE,
                           x_expand = NULL,
-                          x_labels = waiver(),
+                          x_labels = NULL,
                           x_na = TRUE,
                           x_pretty_n = 5,
                           x_rev = FALSE,
@@ -712,6 +716,7 @@ gg_line_facet <- function(data,
       x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = FALSE)
       x_limits <- c(min(x_breaks), max(x_breaks))
       if(is.null(x_expand)) x_expand <- c(0, 0)
+      if(is.null(x_labels)) x_labels <- waiver()
     }
     
     if (is.numeric(x_var_vctr)) {
@@ -747,6 +752,7 @@ gg_line_facet <- function(data,
     }
     else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
       if(is.null(x_expand)) x_expand <- waiver()
+      if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
       
       plot <- plot +
         scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -884,7 +890,7 @@ gg_line_col_facet <- function(data,
                               subtitle_wrap = 80,
                               x_balance = FALSE,
                               x_expand = NULL,
-                              x_labels = waiver(),
+                              x_labels = NULL,
                               x_na = TRUE,
                               x_pretty_n = 5,
                               x_rev = FALSE,
@@ -1007,6 +1013,7 @@ gg_line_col_facet <- function(data,
       x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = FALSE)
       x_limits <- c(min(x_breaks), max(x_breaks))
       if(is.null(x_expand)) x_expand <- c(0, 0)
+      if(is.null(x_labels)) x_labels <- waiver()
     }
     
     if (is.numeric(x_var_vctr)) {
@@ -1042,6 +1049,7 @@ gg_line_col_facet <- function(data,
     }
     else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
       if(is.null(x_expand)) x_expand <- waiver()
+      if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
       
       plot <- plot +
         scale_x_discrete(expand = x_expand, labels = x_labels)