+++ R/gg_point.R
@@ -58,7 +58,7 @@ gg_point <- function(data,
                      subtitle_wrap = 80,
                      x_balance = FALSE,
                      x_expand = NULL,
-                     x_labels = waiver(),
+                     x_labels = NULL,
                      x_pretty_n = 6,
                      x_na = TRUE,
                      x_rev = FALSE,
@@ -143,6 +143,7 @@ gg_point <- function(data,
     x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_breaks), max(x_breaks))
     if(is.null(x_expand)) x_expand <- c(0, 0)
+    if(is.null(x_labels)) x_labels <- waiver()
     
     if(mobile == TRUE) {
       x_breaks <- x_limits
@@ -184,7 +185,8 @@ gg_point <- function(data,
   }
   else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
-    
+    if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
+
     plot <- plot +
       scale_x_discrete(expand = x_expand, labels = x_labels)
   }
@@ -280,7 +282,8 @@ gg_point <- function(data,
 #' @param y_zero For a numeric y variable, TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.  
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles.
-#' @param col_labels_dp Select the appropriate number of decimal places for numeric variable auto legend labels. Defaults to 1.
+#' @param col_labels A function to apply to modify the col_labels. If NULL, categorical variable labels are converted to sentence case, and numeric variable labels to pretty labels with an internal function with col_labels_dp decimal places.   
+#' @param col_labels_dp The number of decimal places for NULL numeric variable default labels. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
@@ -312,7 +315,7 @@ gg_point_col <- function(data,
                          pal_rev = FALSE,
                          x_balance = FALSE,
                          x_expand = NULL,
-                         x_labels = waiver(),
+                         x_labels = NULL,
                          x_na = TRUE,
                          x_pretty_n = 6,
                          x_rev = FALSE,
@@ -334,9 +337,10 @@ gg_point_col <- function(data,
                          col_title = NULL,
                          caption = NULL,
                          col_cuts = NULL,
+                         col_labels = NULL,
+                         col_labels_dp = NULL,
                          col_legend_ncol = NULL,
                          col_legend_nrow = NULL,
-                         col_labels_dp = 1,
                          col_method = NULL,
                          col_na = TRUE,
                          font_family = "",
@@ -409,11 +413,7 @@ gg_point_col <- function(data,
       }  
       col_cuts <- stats::quantile(col_var_vctr, probs = col_cuts, na.rm = TRUE)
       if (anyDuplicated(col_cuts) > 0) stop("col_cuts do not provide unique breaks")
-      
-      data <- data %>% 
-        dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
-      
-      col_labels <- sv_labels_from_cuts(col_cuts, col_labels_dp)
+      if(is.null(col_labels_dp)) col_labels_dp <- 1
     }
     else if (col_method == "bin") {
       if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
@@ -421,12 +421,14 @@ gg_point_col <- function(data,
         if (!(dplyr::first(col_cuts) %in% c(0,-Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
         if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
       })
-      
-      data <- data %>% 
-        dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
-      
-      col_labels <- sv_labels_from_cuts(col_cuts, col_labels_dp)
+      if(is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
     }
+    
+    data <- data %>% 
+      dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
+    
+    if(is.null(col_labels)) col_labels <- sv_numeric_bin_labels(col_cuts, col_labels_dp)
+    
     n_col <- length(col_cuts) - 1
     if (is.null(pal)) pal <- sv_pal(n_col)
     else pal <- pal[1:n_col]
@@ -440,7 +442,7 @@ gg_point_col <- function(data,
     if (is.null(pal)) pal <- sv_pal(n_col)
     else pal <- pal[1:n_col]
     
-    col_labels <- waiver()
+    if(is.null(col_labels)) col_labels <- function(x) snakecase::to_sentence_case(x)
   }
   
   if (pal_rev == TRUE) pal <- rev(pal)
@@ -465,6 +467,7 @@ gg_point_col <- function(data,
     x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = mobile)
     x_limits <- c(min(x_breaks), max(x_breaks))
     if(is.null(x_expand)) x_expand <- c(0, 0)
+    if(is.null(x_labels)) x_labels <- waiver()
     
     if(mobile == TRUE) {
       x_breaks <- x_limits
@@ -506,6 +509,7 @@ gg_point_col <- function(data,
   }
   else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
     if(is.null(x_expand)) x_expand <- waiver()
+    if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
     
     plot <- plot +
       scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -636,7 +640,7 @@ gg_point_facet <- function(data,
                            subtitle_wrap = 80,
                            x_balance = FALSE,
                            x_expand = NULL,
-                           x_labels = waiver(),
+                           x_labels = NULL,
                            x_na = TRUE,
                            x_pretty_n = 5,
                            x_rev = FALSE,
@@ -732,6 +736,7 @@ gg_point_facet <- function(data,
       x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = FALSE)
       x_limits <- c(min(x_breaks), max(x_breaks))
       if(is.null(x_expand)) x_expand <- c(0, 0)
+      if(is.null(x_labels)) x_labels <- waiver()
     }
     
     if (is.numeric(x_var_vctr)) {
@@ -768,6 +773,7 @@ gg_point_facet <- function(data,
     }
     else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
       if(is.null(x_expand)) x_expand <- waiver()
+      if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
       
       plot <- plot +
         scale_x_discrete(expand = x_expand, labels = x_labels)
@@ -863,7 +869,8 @@ gg_point_facet <- function(data,
 #' @param y_zero For a numeric y variable, TRUE or FALSE of whether the minimum of the y scale is zero. Defaults to TRUE.
 #' @param y_zero_line For a numeric y variable, TRUE or FALSE whether to add a zero reference line to the y scale. Defaults to TRUE if there are positive and negative values in y_var. Otherwise defaults to FALSE.  
 #' @param col_cuts A vector of cuts to colour a numeric variable. If "bin" is selected, the first number in the vector should be either -Inf or 0, and the final number Inf. If "quantile" is selected, the first number in the vector should be 0 and the final number should be 1. Defaults to quartiles. 
-#' @param col_labels_dp Select the appropriate number of decimal places for numeric variable auto legend labels. Defaults to 1.
+#' @param col_labels A function to apply to modify the col_labels. If NULL, categorical variable labels are converted to sentence case, and numeric variable labels to pretty labels with an internal function with col_labels_dp decimal places.   
+#' @param col_labels_dp The number of decimal places for NULL numeric variable default labels. Defaults to 1 for "quantile" col_method, and the lowest dp within the col_cuts vector for "bin".
 #' @param col_legend_ncol The number of columns in the legend. 
 #' @param col_legend_nrow The number of rows in the legend.
 #' @param col_method The method of colouring features, either "bin", "quantile" or "category." If numeric, defaults to "quantile".
@@ -906,7 +913,7 @@ gg_point_col_facet <-
            subtitle_wrap = 80,
            x_balance = FALSE,
            x_expand = NULL,
-           x_labels = waiver(),
+           x_labels = NULL,
            x_na = TRUE,
            x_pretty_n = 5,
            x_rev = FALSE,
@@ -926,7 +933,8 @@ gg_point_col_facet <-
            y_zero = FALSE,
            y_zero_line = NULL,
            col_cuts = NULL,
-           col_labels_dp = 1,
+           col_labels = NULL,
+           col_labels_dp = NULL,
            col_legend_ncol = NULL,
            col_legend_nrow = NULL,
            col_method = NULL,
@@ -1009,11 +1017,7 @@ gg_point_col_facet <-
         }  
         col_cuts <- stats::quantile(col_var_vctr, probs = col_cuts, na.rm = TRUE)
         if (anyDuplicated(col_cuts) > 0) stop("col_cuts do not provide unique breaks")
-        
-        data <- data %>% 
-          dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
-        
-        col_labels <- sv_labels_from_cuts(col_cuts, col_labels_dp)
+        if(is.null(col_labels_dp)) col_labels_dp <- 1
       }
       else if (col_method == "bin") {
         if (is.null(col_cuts)) col_cuts <- pretty(col_var_vctr)
@@ -1021,12 +1025,14 @@ gg_point_col_facet <-
           if (!(dplyr::first(col_cuts) %in% c(0,-Inf))) warning("The first element of the col_cuts vector should generally be 0 (or -Inf if there are negative values)")
           if (dplyr::last(col_cuts) != Inf) warning("The last element of the col_cuts vector should generally be Inf")
         })
-        
-        data <- data %>% 
-          dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
-        
-        col_labels <- sv_labels_from_cuts(col_cuts, col_labels_dp)
+        if(is.null(col_labels_dp)) col_labels_dp <- sv_max_dp(col_cuts)
       }
+      
+      data <- data %>% 
+        dplyr::mutate(dplyr::across(!!col_var, ~cut(.x, col_cuts, right = FALSE, include.lowest = TRUE)))
+      
+      if(is.null(col_labels)) col_labels <- sv_numeric_bin_labels(col_cuts, col_labels_dp)
+      
       n_col <- length(col_cuts) - 1
       if (is.null(pal)) pal <- sv_pal(n_col)
       else pal <- pal[1:n_col]
@@ -1040,7 +1046,7 @@ gg_point_col_facet <-
       if (is.null(pal)) pal <- sv_pal(n_col)
       else pal <- pal[1:n_col]
       
-      col_labels <- waiver()
+      if(is.null(col_labels)) col_labels <- function(x) snakecase::to_sentence_case(x)
     }
     
     if (pal_rev == TRUE) pal <- rev(pal)
@@ -1064,6 +1070,7 @@ gg_point_col_facet <-
         x_breaks <- sv_numeric_breaks_h(x_var_vctr, balance = x_balance, pretty_n = x_pretty_n, trans = "identity", zero = x_zero, mobile = FALSE)
         x_limits <- c(min(x_breaks), max(x_breaks))
         if(is.null(x_expand)) x_expand <- c(0, 0)
+        if(is.null(x_labels)) x_labels <- waiver()
       }
       
       if (is.numeric(x_var_vctr)) {
@@ -1100,6 +1107,7 @@ gg_point_col_facet <-
       }
       else if (is.character(x_var_vctr) | is.factor(x_var_vctr) | is.logical(x_var_vctr)){
         if(is.null(x_expand)) x_expand <- waiver()
+        if(is.null(x_labels)) x_labels <- function(x) snakecase::to_sentence_case(x)
         
         plot <- plot +
           scale_x_discrete(expand = x_expand, labels = x_labels)